requirements.txt
asgiref==3.9.1
Django==5.2.6
django-ckeditor==6.7.3
django-cleanup==9.0.0
django-environ==0.12.0
django-js-asset==3.1.2
django-phonenumber-field==8.3.0
django-query-counter==0.4.1
phonenumbers==9.0.16
pillow==11.3.0
psycopg==3.2.11
sqlparse==0.5.3
tabulate==0.9.0
typing_extensions==4.15.0
